Propose alternate team terminology
Hey all. I started reading through slides and am pretty hung up on our
team descriptions and roles. It doesn't map for me to the titles and
roles we have. Thoughts on what I've proposed?

A feature team, as we've been using the term in Delivery land is a set
of engineers working on a feature (!). Those engineers are part of
product engineering team that has all of the roles (Product, UX, Eng).